+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+-------------------------
+##### Fixed
+- No longer working as CommonJS module.
+
 [1.3.0] - 2016-04-26
 -------------------------
 ##### Added